Interstroke unconditional stratified analysis for predictor

Interstroke unconditional stratified analysis for predictor

is_figure_unconditional_stratified_subgroups(
  interstroke_data,
  predictor,
  subgroups,
  model,
  limits = NULL,
  p_sig = NULL,
  breaks = NULL,
  text_size = NULL,
  drop_columns = NULL,
  include_overall = TRUE,
  long_title = TRUE
)

Arguments

interstroke_data
Project specific Interstroke dataset.

predictor
Interstroke predictor of interest for stratified analysis.

subgroups
A list of subgroups for stratified analysis.

model
Model formula for analysis. Predictor should be first independant variable.

limits
**Optional** Limit the x axis in the format c(xmin, xmax).

p_sig
**Optional** Set the level of significance to hide non significant values. Default displays all p values <1.

breaks
**Optional** Set the sequence for breaks in the format seq(xmin, xmax, interval)

text_size
**Optional** Assign value for text size.

drop_columns
**Optional** Drop population and subgroup columns suitable for joint plots.

long_title
**Optional** Set to false to just display predictor as title

Value

A dataframe suitable for Interstroke forest ggplot function.

Examples

is_figure_unconditional_stratified_subgroups(interstroke_data=interstroke_bmi, predictor=whr, subgroups=c("esex","smoking","alcohfreqcat","antihypertpre","cardiacrfcat"), model=case_num~whr+esex+smoking)

#> Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels